balakai/requirements.txt
Django==4.2
djangorestframework==3.15
psycopg2-binary==2.9.9
drf-spectacular==0.27.2
scikit-learn==1.5.1
django-environ==0.11.2
django>=4.2,<5.0